bpf: simplify whitelisting and test it
Be more explicit about how we whitelist. Do not guess if we are
forwarding to tunnel or not. Whitelist both sides if we forward to
tunnel as we go through a single program and policy only.

Added test to check whilisting in various scenatious including Nodeports
forwarded as well as handled locally.