bpf: NodePorts on multi-homed nodes
When a nodeport traffic arrives on one NIC and is forwarded to another
node through another NIC, we need to make sure we use the right source
IP as the next node must be able to route the response back to the
ingress node.

Tests do not cover DSR, but nothing really changes for DSR as long as
the next node has a route to the original client. DSR may not make much
sense in this setup though.

    192.168.20.1              +----------|---------+
         |                    |          |         |
         v                    |          |         V
       eth20                 eth0        |       eth0
     10.0.0.20:30333 --> felixes[1].IP   |   felixes[0].IP
                                         |        |
                                         |        V
                                         |     caliXYZ
                                         |    w[0][0].IP:8055
                                         |
                   node1                 |      node0