bpf: do not accept tunnel from non-node src IP
Prevent accepting decaped packet if sent from a src IP that is not a
known node IP. We fail to create the necessary conntrack entry and that
drop the packet.

We defer the check to conntrack creation as if done earlier, we would
need to do a routing lookup for every packet.