Program blackhole routes for full rejectcidrs to avoid route loops
If nodes have a default route to a Tor which has been advertised a service IP range,
any packets to an IP:port in the service range that doesn't correspond to a service
will route loop to the ToR until hitting the TTL.  This is bad.  Prevent this by
programming a blackhole route for the service ClusterIPs and ExternalIPs.