Pin TEMPEST_BRANCH=23.0.0
Surprisingly this isn't implied by DEVSTACK_BRANCH=stable/rocky, and
needs an independent setting.

Apparently Tempest releases are different from other OpenStack
components.  We have been using master, but we need 23.0.0 as this is
the latest release that supports (a) Rocky and (b) Python 2.  See
https://docs.openstack.org/releasenotes/tempest/v23.0.0.html for more
on this.

Without this pinning, we now (I guess since the release of Ussuri
yesterday) get this error in CI:

    Obtaining file:///opt/stack/tempest
    tempest requires Python '>=3.6' but the running Python is 2.7.17