bpf: HOST_IP represents the same node ip on all host ifaces
Instead of having the iface ip hardwired in the attached tc program, we
hardwire the single node IP that is used to pass packets within the
cluster (the NP vxlan tunnel) and to respond with ICMP errors.

This allows, for instance, for scenarious with multiple NICs that share
the same IP, but the IP is not assinged to any of them, but is
represented by local routes.