Adapt networking-calico code for Python 3 instead of Python 2
Mostly standard changes for converting Python 2 code to Python 3; but
the following points are worth commenting on further.

The etcd3gw package passes all string-like data as Python 3 strings,
except for keys and values that it passes and expects as bytes.  (I
guess this is an unintended consequence of its code for converting
keys and values to/from base64; but no matter.)  Hence our product
code is adapted to call .decode() on those args, on receipt, and to
call .encode() for those args when calling etcd3gw.  The default
encoding for those methods is UTF-8, which is what we want.  Also our
test code that mocks etcd3gw is adapted to use .encode() when sending,
and .decode() on receipt, so that the test code mocks behave in the
same way as the real etcd3gw.

UUID() in Python 3 no longer has a get_hex() method, so we use the hex
attribute as we should have been doing anyway.

In test_snapshot, added mocking for eventlet.spawn so as to avoid
confusion in the test output from _cancel_watch_if_broken running.