Rework to remove race handling service advertisement config
Given a sequence of rapid updates to the config for
<bgpconfig>.spec.serviceExternalIPs and
<bgpconfig>.spec.serviceClusterIPs, it was possible for the goroutines
handling each update to race with each other, and

- at least, end up in a state where the external and cluster IPs
programmed in confd did not match the final datastore state

- hypothetically, cause other bad effects, from similar goroutines
racing against each other.

To avoid that, this rework keeps the handling of external and cluster
IPs updates entirely within the Calico client object; this means
updates can be handled synchronously without involving the lock of the
separate RouteGenerator object.  It also means that the RouteGenerator
is now responsible only for node-specific routes; this was nearly the
case before, but is now 100% the case.