bpf: generic conntrack Scanner
We scan the conntrack table for various reasons and since each scanner
needs to be split per entry type, it is complicated to write it in a
single function. This allows us to write the functions independently,
configure them independently and execute then in a delete safe manner as
scanners do remove entries.

The API does not allow to delete a random entry, but it allows getting a
random entry if that is neede to evaluate the current entry.