Add Tip to Global Network Policy Section (#3460)
* Add Tip to Global Network Policy Section

When I first started working with Calico, I ran into an issue where new network policies I was creating were not behaving as they should. I could not find any network policy that was causing this, but later found out that there was a lingering Global Network Policy that was behind the issue. This tip could have saved me the time I spent trying to address this issue.

* Adjust upper/lowercasing

Co-authored-by: Casey Davenport <caseydavenport@users.noreply.github.com>