Add BPF dataplane support for Wireguard
We need a new BPF program variant for Wireguard interfaces, because a
Wireguard device sees an SKB with the IP header at offset 0, in both
directions.  (aka It's an L3 device.)

For now we prevent direct redirect from the Wireguard ingress problem,
because that would need adding back an Ethernet header.  There will be
follow-on work to try to do that.