bpf-apache: avoid including libbpf.h for BPF-prog
The header file bpf/libbpf.h is meant for userspace programs using libbpf,
and not for BPF-programs (which BPF-bytecode is loaded into kernel).

The problem is e.g. bpf/libbpf.h will include stdio.h, which doesn't make
sense for a BPF-program.

If compiled with -target bpf and not having libc 32-bit stubs installed then
this will result in the following compile error:

  $ make
  In file included from sockops.c:16:
  In file included from ./sockops.h:15:
  In file included from ./bpf.h:19:
  In file included from /usr/local/include/bpf/libbpf.h:24:
  In file included from /usr/include/stdio.h:27:
  In file included from /usr/include/bits/libc-header-start.h:33:
  In file included from /usr/include/features.h:474:
  /usr/include/gnu/stubs.h:7:11: fatal error: 'gnu/stubs-32.h' file not found
  #  include <gnu/stubs-32.h>
             ^~~~~~~~~~~~~~~~
  1 error generated.
  make: *** [Makefile:53: sockops.d] Error 1

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>