Include optional configmap to override k8s endpoint
The config map should contain these env vars:

KUBERNETES_SERVICE_HOST
KUBERNETES_SERVICE_PORT

This is (only) required when enabling eBPF dataplance as kube-proxy
needs to be turned off and without kube-proxy, the kubernetes services
cannot be used.

If the map is not set, the env vars are not set either, so this does not
have any implication for non-bpf setups.