bpf/proxy: Conntrack cleaner runns immediately after sync up
Immediatally after the state of proxy is synced up and the fresh state
is applied, start the conntrack cleaner. It makes its first sweep
immediately. It needs to be after the first sync up to avoid breaking
connections that may look like towards stale backends as the internal
state was not updated yet.