Always apply BPF programs to workload interfaces
Rather than applying programs only when there's an associated
endpoint, apply programs to all workload interfaces.  This closes
a race between removing the program and removig it from the iptables
allow list that polices unknown endpoints.

As a side benefit, it means we get a program ready on a newly-created
interface, ready to revcieve policy when the endpoint is created.