bpf/proxy: ignore bpf deletes ENOENT errors
When trying to delete something that does not exist may be an issue, but
should not stop the syncer from continuing as the result is that the key
does not exist anyway.

There is at least one genuine reason why this happens - when LB source
ranges changed. For simplicity of the code, the service is completely
cleaned up before applying changes, while it is nbeing deleted once
again when cleaning up stale derived services.

This may also happend when either bug or external delete modified the
tables. Therefore we do log it when debug is on.