bpf: use state from map, free up stack
We are running out of space on stack, especially when DEBUG is enabled.
Therefore this change frees up 60 bytes of stack space (limit is 512)
move the jump state away from the stack and getting it from the map
where we are going to update it anyway if we needed to go through
policy.

This adds a lookup in CPU local array (fast) for each packet, removes
updates for packets that go through policy.