bpf: harden accept_local using iptables rpfilter
We need accept_local for traffic that local IP when we turn the packet
around from ingress to egress side. The primary reason is ICMP reponses,
however, it is also used the vxlan NodePort tunnel.

The NodePort tunnel could avoid that, however, that would require
relaxing the RPF check when an overlay is used for pod-pod networking as
we decap the packets on external phys interfaces, but the best route is
through the tunnel interface.

It is easier to harden accept_local then to relax rpf for selected
packet in iptables.