bpf: enforce RPF check for possible spoofed TCP
TCP sets CALI_CT_ESTABLISHED_BYPASS which avoid parsing the packet by
the second program on established connections. However, if we detected
change in where the packets come from - possible spoof, disable the
bypass. As the packet is forced through the kernel's RPF check, avoiding
the bypass updates the ifindex if the packet passes through.