Make DHCP agent restart more reliable
It has been reported to us that in some circumstances (crashes caused
by other bugs) the Calico DHCP agent does not always restart
automatically, and has to be restarted manually with `systemctl
restart calico-dhcp-agent`.  We think this is because systemd does not
interpret some crash circumstances as matching `Restart=on-failure`;
changing that setting to `Restart=always` should be more reliable.  In
practice the Calico DHCP agent should always be running, unless it has
been explicitly stopped with `systemctl stop`, and `Restart=always` is
the correct setting for that desired behaviour.

We also set `RestartSec=3s` to slow down a bit if the DHCP agent is
restarting repeatedly.  Compared to the default of 100ms, this allow a
little time for contributing conditions to clear (possibly), but is
still plenty fast enough for practical DHCP service, given that DHCP
clients will retry.

Finally, remove `KillMode=process`, as this setting is now not
recommended (see [1] and [2]) - and I'm not aware that we ever had a
specific reason for choosing the `process` setting.  Removing the
setting means that `KillMode` will default to `control-group`.

[1] Current systemd.kill manpage:
https://www.freedesktop.org/software/systemd/man/systemd.kill.html

[2] Commit that added the non-recommendation:
https://github.com/systemd/systemd/commit/9b52e0d81af9c0fbdfc8686eecacc334f8d60aa8#diff-d050734015b4e77c2348f3ac3799a12e