bpf: skip FIB for WEP egress leaving the cluster
When traffic from a pod is leaving the cluster, we may want to decide
to skip FIB and thus force it to go through the host IP stack and IP
tables.

For instance, in EKS, the AWS CNI programs nat outgoing rules and we
want the traffic to hit them.

We skip the fib when the source IP of the workload is not in any IP
pool, which suggests that a non-calico CNI is in use.