Add a file lock to serialise IPAM allocations.
Acquire a best-effort host-wide lock to prevent multiple copies of the CNI plugin trying to assign
concurrently. AutoAssign is concurrency safe already but serialising the CNI plugins means that
we only attempt one IPAM claim at a time on the host's active IPAM block.  This reduces the load
on the API server by a factor of the number of concurrent requests.