mech_calico: Ignore pre-live-migration messages
As of OpenStack Newton, an RPC call push was added to pre-
live-migration steps. Reference PRs:
https://review.opendev.org/c/openstack/nova/+/275073
https://review.openstack.org/#/c/260738/

This commit adds a guard to `mech_calico`'s port updating
logic so that it effectively ignores the pre-live-migration
notification referenced above.  Without this guard,
`mech_calico` mistakeningly thinks the port is becoming
unbound/disabled (and later enabled again) as part of the
pre-live-migration notifications and flaps the VM's port
accordingly.

In short, the addition of the guard is required to avoid
an unconditional, intermittent loss of connectivity for VMs
during the live-migration process.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>