Disable Semaphore's copy-on-write docker mount.
The copy-on-write docker mount is intended to provide a cache
of useful images but it doesn't cover any images that we use.

In addition, as we churn through docker containers in the FV tests
the copy-on-write FS grows and grows until it fills the root FS.
It's actually better for us to simly write to the root FS directly
since deleted files on the root FS can actually be reclaimed.