bpf/conntrack: scanners IterationEnd called in reverse order
To minimize the danger of scanners deadlocking, call the IterationEnd()
in a reverse order than the IterationStart so that if scanners acquire
locks and resources, they release them in the LIFO way.

Atm the only scanner that acquires a lock is the StaleNATScanner, which
calls into the kubeproxy, which internally grabs a lock.