bpf: EKS workaround iptables rule for BPFExtToServiceMark
EKS tries to restore the 0x80 mark from connection. This removes the
BPFExtToServiceMark we set on egress from pods if we do not set it on
the connection in host namespace.

Preferably, EKS would restore the mark only if the mark was set and
would leave the mark untouched otherwise.