Stop ARP traffic being dropped due to RPF check
ARP request packets were being dropped on the receiving node due to the
RPF check. This was because the incoming packets arrived on eth0, but
the RPF check determined that as the packets were not marked with the
wireguard fwmark, the return route would be through the wireguard
interface.

To fix this issue, iptables rules were introduced to mark all packets
which were not from the lo, wireguard, or cali+ interfaces with the
wireguard fwmark. This has the affect that these packets will then not
be routed through the wireguard interface and will pass the RPF
check.