Update OpenShift operator RBAC for certification
The SCC rule in the operator's cluster role is only for operator certification purposes.
The operator normally runs in a namespace with openshift.io/run-level=0 so SCC is bypassed.
In certification tests, the operator is run in a normal namespace so this
rule is needed for host networking and hostPath volume access.