Add an extra grace period for new connections
We've seen felix get cyclicly booted from the server, presumably
because sending the snapshot took an unexpectedly long time due
to CPU or netowrk contention.

Add an extra grace period for newly-connected clients so that
they get some time to catch up after the snapshot is sent. Hopefully,
that will solve the problem outright but it also has the side effect
of giving the new felix some time to update the dataplane before we
kill the Typha connection.  This ensures that it makes progress
even if it's restarting.