etcdv3.py: Do not subclass etcd3gw.watch.Watcher
`etcd3gw>=0.2.2` incorporates the fix that the subclassing
in this module itself was doing.  However, as of new
`etcd3gw`, this now results in an attempt to close the same
file descriptor twice when `stop()` is invoked, which
itself results in an exception that can result in log
spew, as seen below:

```
(WARNING) etcdv3: Failed to force-close etcd3gw's watch socket: AttributeError("'HTTPResponse' object has no attribute '_fileno'",)
```

Pin `etcd3gw` in the requirements so that the version with
the fix is used, and remove the fix incorporated in this
module.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>