Correct unit testing for a delete / update race
We have a unit test for when a port is updated and then quickly
deleted, and the update and delete callbacks are handled out of order
by different Neutron server threads, but it wasn't correctly modelling
what the update callback would see in that case, namely a PortNotFound
exception when trying to re-read the port data.  This commit corrects
that.