Corrections to update code for migration and delete/update race cases
The previous non-UT commit removed `_migration_step`, but an important
detail of `_migration_step` was that it allowed for the new
WorkloadEndpoint (on the new host) not to exist yet, whereas other
branches of `update_port_postcommit` require that the WorkloadEndpoint
already exists.  Here we fix that by introducing an
`endpoint_should_already_exist` variable, that also replaces
multiple calls to `port_bound(original)`.

Also add handling for PortNotFound exception, which is expected in
case of a delete/update race.  (See previous commit for more detail
about that.)

Finally restructure the various cases to be hopefully a bit clearer.