Fix slow performance of updating a namespace.
Previously we assumed that namespace updates were rare so
we did a full scan of the endpoint index to find any endpoints
matching a particular namespace.  That assumption doesn't hold
on clusters with many namespaces because we need to process
the namespaces at start of day.

Add an index that tracks the enpdoints that have a given
parent (i.e. namespace) so that we can do a fast lookup.

Add benchmarks to verify the change.