bpf: source port collision resolution
If there is a collision on the source due to forwarding via multiple
nodes, we pick a new source port, we remember the source port in
conntrack and we SNAT the port on the way towards the pod and we
translate it back for the return traffic. That port is not seen outside
of the node.

We remember the new port in the FWD entry an we also remember the
original source port in the reverse entry. We use currenty unsed space
in the entries, hence there is no need to version the conntrack table.