bpf: test random port collision
After we have a collision, we pick a random port. That may lead to
another collision with a previously selected random port. In such a case
we drop the packet and we let the net protocol to resolve this. TCP
retries a few times and that should resolve the secondary conflict.

Note that these conflicts should be rare to start with, even more rare
if we there is a secondary conflict.