bpf/proxy: exclude local workloads from NodePortRemotes
Fixes the problem, where bogus entries were written to NAT table due to
the fact that local workloads do not have NextHop as an IP, but as the
index of the cali interface. That lead to programming bogus entries like
3.0.0.0, 4.0.0.0 etc in the NAT table.

As a side effect, those entries each had just one backend (the pod with
the cali iface) and due to another bug overwrote the backends for the
cleaner that thought that there is just a single backend which leads to
stalls if there is more then a single local backend
projectcalico/calico#4431. This PR fixes only programming the NAT table
and does not address the cleaner.

To work around the fact that CTLB does not resolve nodeports at the
node, but immediately at connect time, we need to know if there are any
local pods at the destination node if ExternalTrafficPolicy is set to
Local. Otherwise we would not be compliant with legacy kube-proxy if we
make connetion to a non-local pod of the destination node.

Note that CTLB would make such a connection with source preservation and
now extra hop.