Avoid error when bpftool returns empty JSON output at start of day
A Calico/OpenStack customer saw this:

    [ERROR][61494] felix/connecttime.go 63: BPF program list not json. error=unexpected end of JSON input output="\n"

Customer wasn't trying to run in BPF mode, but this can occur at start
of day even when not BPF-enabled, as part of cleaning up old BPF
state.

We presume they have a version of bpftool, on the host, that behaves
differently from the one that we include in a calico-node image;
specifically, such that `bpftool -j -p cgroup show <cgroupPath>` gives
output `\n` and exit status 0; and that trips us up.  ￼