bpf/ctlb: fail connect if there is no backend
We may hit a race when backends are not programmed yet and if we let
TCP connect succeed, a SYN from pod will hit NAT in tc program, which
does not work in all cases. It would not work for host at all until the
TCP session expired and application retries.

This also makes it aligned with tc programs that generate ICMP response
host-unreachable in such a situation.

Connect returns EPERM in this case and application may retry immediately.