windows: replace install-calico-windows.ps1 template (#5166)
* windows: replace install-calico-windows.ps1 template

The node repo is now the source for the Calico for Windows installation
script. Add targets to update the script here and to check that we have
the correct release version of the script checked in for release.

This commit replaces the template with the latest version of the script
from node.

* Review markups

* Review markups

Make it clearer that the master version of the install script needs the
user to provide the installation zip file.