Return rule conversion errors from K8sNetworkPolicyToCalico()
Add ErrorPolicyConversion which contains a slice of ErrorPolicyConversionRules,
so that K8sNetworkPolicyToCalico() in libcalico-go/lib/backend/k8s/conversion/conversion.go
returns all rules that it could not convert (this will be used to implement a policy
conversion tool in calicoctl).

Make current uses of K8sNetworkPolicyToCalico() ignore the error if it is
ErrorPolicyConversion (as they were currently functioning without having these returned anyway).