bpf: enforce policy on SYN packets
In case policy has changed and there is an existing conntrack entry for
previously allewed connection, a new connection with the same tuple
would hit the CT entry and would be let through even if the current
policy does not let it anymore.

This is a corner case, but a client can enforce its source port as our
named ports FV tests do.