fv: workaround for conntrac issue in BPF in named ports tests
When a connection with both fixed src/dsr ports makes is through, and
when policy changes then, next connections are exempt from policy and
the test may fail.

This change cleans up conntrack between retries to avoid that. This is
purely a workaround and should be reverted once there is a proper fix!