felix/bpf: use port SNAT port when cleaning conntrack
When we clean conntrack entries for connections that connect to
no-longer existing backend, use the port SNAT port to identify which
side of the the forward CT entry is the client. Then we can identify
which is the service and which is the backend.

Without using the port SNAT port, the client IP matches, but the port
does not as the port in the key is the original port.