Pin client-go to version with fix for #5173
When KDD does a paginated list with a revision, it sets
ResourceVersionMatch as well as ResourceVersion.  The client-go
pager was sendign the ResourceVersionMatch parameter on continuation
pages resulting in an error.  This caused the watcher to consistently
fail after hitting that condition.

Pin to a version of client-go that unsets ResourceVersionMatch
on continuation pages.