felix/bpf: set DNAT when enforcing policy on SYN
Unlike NAT lookup, CT lookup does not set the DNAT in state, which is
fixed only after policy. Therefore we need to set it properly if we want
to enforce policy correctly.

Adds a test that changes the policy and checks that the policy is
enforce on a retransmitted SYN - equal to a fresh new SYN hitting an
existing CT from previous connection.