felix/bpf: disabled CTLB workaround for host and nodeports
Packets from the host are intercepted at the first device they hit. If
it is towards the same host, it is the loopback device, otherwise it is
the main device on the cluster network. At that point we do the DNAT -
select the backed pod for the nodeport. We assume that there is a
default route so it hits a device. Otherwise we cannot do much.

We either use FIB to redirect the packet to an outgoing device (if on lo
or if it should go to a different one than the default) or we redirect
the packet to the bpfin.cali iface to turn the packet around into the
host address namespace that can route the packet. If an overlay is used,
we SNAT the packet so that it has the tunnel's source and we redirect it
to the tunnel.