Add sources to chart so the source of the chart is known
It is always hard to understand an chart without it's source, especially this chart since the values are hardly document (the entire space)
So it is useful to understand where the source code is, which is fairly hard to find right now (i had to ask in slack myself to be pointed here).

I would love to improve this for others to find it an easier way. Thoughts?