Have Felix use IP pools to determine enabled encaps
Mark IpInIpEnabled and VXLANEnabled as deprecated with comments and make them pointers in config_params.go to match FelixConfiguration, to make it possible to distinguish an 'unset' nil value from an explicitly disabled false.

Add Encapsulation struct to config_params.go which contains the calculated IPIPEnabled and VXLANEnabled parameters that represent whether these encaps are effectively enabled or not.

Add EncapsulationResolver to calc graph which will watch the IP pools and send an Encapsulation protobuf message to the dataplane when IP pool encapsulations change. The dataplane will restart Felix if there are IPIP and/or VXLAN changes (so that the encaps are recalculated and applied) when receiving a proto.Encapsulation message in sendMessagesToDataplaneDriver() in daemon.go. Also add support for exiting cleanly on reasonEncapChanged to monitorAndManageShutdown() in daemon.go.

Add EncapsulationCalculator helper struct, used by Run() in daemon.go and by EncapsulationResolver. It calculates whether the encaps should be enabled based on the existing IP pools and FelixConfiguration.

Add support for Encapsulation in the form of OnEncapUpdate() and flushEncapUpdate() to event_sequencer.go.

Add *bool parsing support to param_types.go.

Add encapsulation_resolver_test.go UTs to test EncapsulationResolver and EncapsulationCalculator exhaustively. Add encapsulation support and test cases to calc_graph_fv_test.go.

Update docs accordingly.

Update description for IPIPEnabled and VXLANEnabled in FelixConfig CRDs