WIP: Have Felix use IP pools to determine enabled encaps
Make IpInIpEnabled and VXLANEnabled pointers in config_params.go to match FelixConfiguration, to make it possible to distinguish an 'unset' nil value from an explicitly disabled false.

Add EncapInfo to config_params.go which contains the calculated UseIPIPEncap and UseVXLANEncap parameters as well as the sets of IP pools that have IPIP and VXLAN enabled.

Add a step to Felix startup in daemon.go to calculate if the IPIP and VXLAN encaps should be enabled, based on FelixConfiguration propagated from config_params, and the existing IP pools.

Add PoolEncapManager to calc graph which will watch the IP pools and restart Felix if there are IPIP and/or VXLAN changes (so that the encaps are recalculated and applied).

Add *bool parsing support to param_types.go.

Remove maybeEnableIPIP() and maybeEnableVXLAN() functions, as these shouldn't be changed in FelixConfig anymore.

Explicitly create default FelixConfiguration for flannel migration in kube-controllers/pkg/controllers/flannelmigration/ipam_migrator.go (instead of having IP pool creation do it).

Add cleanUpIPIPAddrs() to int_dataplane.go and call it when IPIP encap is disabled to clean up old addresses from tunl0.

Tests are pending, but I wanted to push this to get eyes on it early on.