felix/bpf: do SNAT w/o CTLB when target is via tunnel
Do not rely on iptables to fix up source when the backend pod is
accessible via tunnel as that makes routing super complicated.

We change only the source IP and we leave the port untouched as that
would only collide if the host was also accessing the same pod directly
(not via a service). If that happens, we we resolve the source port
collision in the same way as we do for nodeports. That is, we pick a
random port from a range that is not used for ephemeral ports. We can
pick multiple times if there is a collision.