felix/bpf: use correct affinity key for unconnected udp
We enforce affinity on unconnected udp when we use ctlb because
otherwise we would make a NAT decision for every packet, possibly
spraying them all over the backends even for a single client.

We were incorrectly initializing the affinity key so that the key was
zeroed and we are not making the actuall lookup.

FV tests that the new key is expected and that the values are consistent.