felix/bpf: good to know which routes are tunneled
Internal routes carry the knowledge whether they go via an overlay
tunnel or not. BPF routes where missing this knowledge.

In the current setup, we assume that when Wireguard is used, any pod is
reached by a tunnel without that being explicit in the internal routes.

This is a prerequisite for being able to forward correctly without CTLB
and to do tunneling in BPF entirely.