bpf/fv: test source port resolution for SNAT
When a host accesses a service without CTLB both via its service IP and
directly, there may be a conflict on the source port after we SNAT the
IP after service resolution. This tests that we can handle the conflict
regardless of whether the coflict was identified at the bpf nat device
or the host interface.