Report health detail
This change surfaces more detail about the internal components, within a given process such as
Felix, that contribute to that process's overall liveness and readiness state.  If a process is
reporting non-live, or non-ready, and it isn't immediately obvious why, this extra detail should
indicate where the problem is.

Here's an example for Felix during startup, when it hasn't yet managed to connect to the datastore:
```
+---------------+---------+----------------+---------------------+--------+
|   COMPONENT   | TIMEOUT |    LIVENESS    |      READINESS      | DETAIL |
+---------------+---------+----------------+---------------------+--------+
| felix-startup | 0s      | reporting live | reporting non-ready |        |
| int_dataplane | 20s     | reporting live | reporting non-ready |        |
+---------------+---------+----------------+---------------------+--------+
```

On the internal health API, this is provided as the `Detail` field of the report returned by
`HealthAggregator.Summary()`.  It's also now possible for individual health reporters to fill in
`Detail` when they make a report, and that would be shown in the "DETAIL" column.  (But no existing
coding sets any per-reporter `Detail` yet, as this commit has only just created the `Detail` field.)