Rework interface monitor
Previously, the iface monitor used "down" to mean two things:

- Interface was previously signalled as "up" and now it has gone  "oper down".
- Interface was previously signalled as "up" and now it has been  deleted.

It never signalled "down" interfaces that had never previously been "up".

This was confusing and resulted in mismatched expectations in the AWS IP manager, which was expecting "down" interfaces to be signalled (so that it could bring them "up").

Standardise on signalling:

- "up" for present, "oper up" interfaces.
- "down" for present, "oper down" interfaces.
- "" for deleted interfaces.