felix/bpf: CALI_SKB_MARK_BYPASS_FWD should skip RPF
When CALI_SKB_MARK_BYPASS_FWD we are forwarding traffic returning from a
tunnel (known) to the client (known).

It is a special case in which skip RPF was wrongly enforced, however, to
skip RPF we cannot set both at the same time. Therefore make the
iptables rules that enforce the strict check skip it.