[RFC] Check active policy only for policy resolver
Currently the policy resolver put all the policies into the policySorter
and scan it when a policy becomes matched or mismatched.

However, we need only put the active policy into the policy sorter and
when a policy match change happens, we don't need scan all the policies,
but only the active one. This should reduce a lot of effort.

A new field is added to keep all the policies.