Enable Multi-arch building with Unified Dockerfile
Now we try to enable a unified Dockerfile to build images
for multiple architectures(here amd64 and arm64), it's now
for the apiserver component here.
With more agreements, the way used in this PR can be
popularized to other Calico components.

To use a single Dockerfile for image building would greatly benefit the
maintenance and development work for us. Here we think for most
cases a unified Dockerfile can work with a unified view.

Signed-off-by: TrevorTaoARM <trevor.tao@arm.com>