felix/bpf: include RPF and wg chains again. (#5342)
Add strict RPF enforcement on all interfaces but allow to control whether it is enforced by
BPFEnforceStrictRPF=true/false config option. That option is set to true by default as it is what most users
want anyway.

This is also a pre-work for not requiring CTLB in some corner cases. In that case we will need to relax RPF
at the devices and enforce it when necessary ourselves.