Enable Native Building for Node Image on arm64
This PR enables the native building for node image
on arm64 platform which fixes the error met and
adds target arch build-arg to enable possible unified
Dockerfile in the future.
The native building for Node image on arm64 is needed
for running the e2e test on arm64.

When building on native arm64 platform(not cross-building
on amd64), the modification to Runit source code is not
needed and must be ignored to avoid building errors.

And we switch to use the same set of Runit src code as
that of amd64 which would be more consistent for both
platforms.

Verified the building both on amd64(crossbuilding)
and arm64.

Signed-off-by: TrevorTaoARM <trevor.tao@arm.com>