set the NODENAME from spec.nodeName
While the install script will auto-determine and set NODENAME based on
environment this may not always be correct (e.g. in AWS, windows
nodes might be using the COMPUTERNAME but the install script will use
the local hostname from the AWS metadata service.

Setting the NODENAME from the spec will always be right.