felix/bpf: always enforce strict RPF on WEP
Felix always enforces strict RPF on workload endpoints. In BPF mode the
BPF programs do that on the first packet using its internal routing
information. However, when there was a conntrack hit on a subsequent
packet, but the packet came from an unexpected interface, it was
deferred to host network stack to make the test. With this change the
check is done completely in BPF, which makes it possible to prevent
spoofing regardless of the workload (cali) interface rp_filter setting.

In addition, this allows us to relax the rp_filter setting that would
interfere with SNATing packets at the workload interface, e.g. when
returning packets on nodeport connections or reversing service towards
host without CTLB, thus simplifying how we handle such packets.