Set the tunnel mtu same as veth mtu for workload programs. Earlier, (#5937)
we were passing packet length to the fib lookup helper which was doing
mtu check and thus failing if the packet length is greater than veth mtu.
Now that we pass 0 to fib lookup, mtu check is skipped. As a result, we
can accomodate 50 bytes more in case of nodeport in the return path.