libcalico: watcherCaches must be marked as started soon
If we want to signal watcherSyncer to flush and stop by closing the
ws.results channel, we must mark the watcherCaches as started at the
same time as starting the syncer thread to synchronize the unwinding
properly. They write to the channel and panic if the channel is closed
while they are writing into it.