Filter BGP everse peerings to same IP version.
When doing an explicit peering between two Calico nodes on IP address,
make sure that we only try to peer over the same IP version.  Previously,
we treated an IP amatch the same as a selector match when calculating
reverse peerings, this meant that the forward peering would be, say
IPv6-only but ht ereverse peering tried to peer over IPv4 and IPv6.