Avoid rebuilding felix image for every FV run.
For a long time, the felix image has been an FV-only test image.
This commit changes that image into a "test environment" image
that doesn't include felix or the test tool binaries themselves.

This means that the image can be built once and then reused for
multiple test runs by mounting in the updated felix binary for each
run.

This required various knock-on changes:

- The non-AMD64 docker images are never used for FVs and had rotted so
  I removed them.  We weren't publishing these anywhere.

- Mounting in the test binaries required moving them to a new directory
  so I put them on the path inside the container and changed all
  instances of "/test-binary-name" to just "test-binary-name".

- Creating an external node was done all over the FVs; I made that code
  common so I only had to fix up one place.

- Use a touchfile to avoid rebuilding the felix image.

- Remove tini from the container image and us "docker --init" instead.
  Downloading tini has been flaky in the past.